change the Shape interface
Now a shape have a partition function that return the variable in one
side and the partition of the type with a shape on the other.
Here the notion of variable is broad, it mean anything that we treat as
variable. In the case of Const it is everything thing except constant
and frozen variables.