AC solution are now Env.t
We change how we handle solution for the AC problem. Before, we were
constructing a set of diophantine solution, converting them into
a set of equation and inserting them into the env. This has several
draw back. First, if a solution of a diophantine system is not valide,
we will still try it potentially many time. Second, we process the
equation of type of each diophantine solution several time.
To fix, now each solution of each diophantine system, is converted
into a set of type equation and partial assignement, then the type
equations are inserted in the system in a copy of the Env. By doing this
we can see early if one diophantine equation is not valide and we
process them only once.
To build a full solution, we need to merge the Env of each part of the
solution, once this is done, we need to turn the partial assignements
into full assignements (Env.commit)