Fix the fact that Shape don't deal with non_arrow_var correctly
Currently, shape does not distinguish non_arrow_var and normal var.
Therefore, an AC solution, could associate a non_arrow_var to an arrow
type. With this fix, if we try to get the representative in such case we
raise an exception which is catch to make this branch fails. A long term
fix will be to add the support for non_arrow_var into the shape