Fix the insertion of non-proper equation.
In the case of a non-proper equation "x = y" such that both variables
have the same value "t". Then, if we introduce the equation "x = y" and
"y = x", this creates a loop and get rejected by the occur_check. The
problem here is that the order of "x" and "y" is defined with respect to
the value associated to it. This commit changes that to compare directly
the variables. This change guaranty that the non-proper equation will
form a DAG and prevent any cycle.
As a side effect, this makes the comparison faster.