Fix bug in the AC solver with respect to Arrow and Non arrow var
Earlier, in the AC solver, we assumed that a non arrow variable could
not be instanciate with an arrow, but the AC solver is building a tuple,
therefore it is valide to instanciate a non arrow variable `v` could with
`int * (int -> string)`. Now Arrow and NonArrow var are not longer
treated differently than the rest.